Move update code out of main
